refactor(image-plugin-redesign): rename new component to image-selection-screen
